+++ demo/car-crashes/crashed/pipeline/ops/testing.py
@@ -15,11 +15,11 @@ from typing import List, Dict
 
 from hypermodel.platform.gcp.services import GooglePlatformServices
 
-from hypermodel.ml.features.categorical import (
+from hypermodel.hml.features.categorical import (
     get_unique_feature_values,
     one_hot_encode,
 )
-from hypermodel.ml.model_container import ModelContainer
+from hypermodel.hml.model_container import ModelContainer
 
 
 from crashed.model_config import crashed_model_container, build_feature_matrix
