+++ src/hyper-model/hypermodel/hml/decorators.py
@@ -1,6 +1,7 @@
 import logging
 import click
 import kfp.dsl as dsl
+from hypermodel.hml.model_app import ModelApp
 
 from typing import List, Dict
 
@@ -11,7 +12,7 @@ class ModelOpWrapper:
         self.name = func.__name__
 
 
-def model_op():
+def op():
     """
     Here we are going to wrap the function that we are trying to execute
     so that we can get meta-data about the function such as its command
@@ -30,12 +31,27 @@ def model_op():
     return _register
 
 
-def model_op_param(*args, **kwargs):
+def option(*args, **kwargs):
     print(f"@model_op_param")
 
     def _register(func):
         print(f"@model_op_param _register: {func}")
         # print(f"model_op._register()")
         return click.option(*args, **kwargs)(func)
+
+    return _register
+
+
+def pipeline(app: ModelApp):
+    def _register(func):
+        print(f"model_pipeline._register: {app.name} for {func.__name__}")
+
+        pipe = app.register_pipeline(func)
+
+        def _func_wrapper(*args, **kwargs):
+            # Execute the function
+            return func(*args, **kwargs)
+
+        return pipe
 
     return _register
