+++ src/hyper-model/hypermodel/hml/features/categorical.py
